Convert Bootstrap Summary Tables to flextable Objects

Converts tables created using boot_summary and censboot_summary to nicely formatted flextable tables.

Usage

summary_to_flextable(summary_table, decimals = 3, conf = "95 % CI")

Arguments

summary_table

A table created using boot_summary or censboot_summary.

decimals

The number of decimals to print for estimates and confidence intervals. The default is 3.

conf

The text at the top of the confidence interval column in the gt table. The default is "95 % CI".

Value

A flextable object.

Examples

# Bootstrap summary of a linear model for mtcars:
model <- lm(mpg ~ hp + vs, data = mtcars)
boot_summary(model, R = 99) |> summary_to_flextable()

 

Estimate

95 % CI

p-value

(Intercept)

26.963

(22.414, 33.997)

<0.01

hp

-0.055

(-0.088, -0.032)

<0.01

vs

2.576

(-1.672, 5.825)

0.31

# Export to Word:
if (FALSE) { # \dontrun{
boot_summary(model, R = 99) |>
   summary_to_flextable() |>
   flextable::save_as_docx(path = "my_table.docx")
} # }